pyproject.toml
[tool.poetry]
name = "m2-hw12"
version = "0.1.0"
description = ""
authors = ["Pelmenoff <Titanfall@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
psycopg2 = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.5.1"}
sqlalchemy = "^2.0.23"
python-jose = "^3.3.0"
uvicorn = "^0.24.0.post1"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
alembic = "^1.12.1"
bcrypt = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
